faceDetector_with_genderRecognition/requirements.txt
mtcnn==0.1.1
deepface==0.0.79
opencv-python==4.7.0.72
streamlit==1.19.0
numpy==1.24.2
Pillow==9.4.0
